requirements.txt
Flask==3.0.3
gunicorn==23.0.0
openai==1.58.1
psycopg2-binary
